style: getFormatDate 함수 Utils에 분리
